requirements.txt
Flask==2.0.3
google-cloud-datastore==1.15.0
google-auth==1.30.0
requests==2.27.1
